fix(advanced-marker): apply marker class when rendering a Pin
Previously, the AdvancedMarker would only apply the className prop to the content when custom html contents were specified, but there are some cases (applying css-animations to markers for example) where we also want the className to apply to markers that just contain a PinElement.

This also contains a rewrite and extension of the AdvancedMarker tests.