docs(examples): fixed and extended marker-clusterer example
The marker-clusterer example included a couple of problems that would
only surface when being extended in certain ways. These were mostly
related to missing memoization of callbacks.

This has continued to cause some problems in the community so I
attempted this version that handles the typical additions like an
infowindow and dynamic marker updates without problems and in a mostly
idomatic react way.