feat: add hover events and anchor points to advanced markers (#472)
This is a significant change to how the AdvancedMarker component works internally and adds two major new features to advanced markers:

 - `Event Handling`: in addition to the click and drag events, the Advanced marker now handles mouseenter and mouseleave events correctly, independent of being used with a Pin or with custom html content

- `Anchoring`: the AdvancedMarker component now supports configurable anchor-points, to make it easier to create different kinds of markers. This also new properly works together with the infowindows.

This change required us to change the DOM structure created by the AdvancedMarker by add an additional div around the content. If you’ve been using very specific selectors to style the content of an AdvancedMarker, this might require you to update them accordingly.